createCampaignTest

+++ src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 DROP ALL OBJECTS;
 
 CREATE TABLE IF NOT EXISTS CAMPAIGNS(
-                                        id         INT AUTO_INCREMENT PRIMARY KEY,
+                                        id         BIGINT AUTO_INCREMENT PRIMARY KEY,
                                         name       VARCHAR(100) NOT NULL,
                                         category   INT,
                                         bid        DOUBLE       NOT NULL,
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS CAMPAIGNS(
 
 CREATE TABLE IF NOT EXISTS PRODUCTS
 (
-    id         INT AUTO_INCREMENT PRIMARY KEY,
+    id         BIGINT AUTO_INCREMENT PRIMARY KEY,
     name       VARCHAR(100) NOT NULL,
     start_date DATE,
     bid        DOUBLE       NOT NULL,
